// src/common/enums/loan-application-status.enum.ts
export enum LoanApplicationStatus {
  DRAFT = 'DRAFT',
  SUBMITTED = 'SUBMITTED',
  CALL_VERIFIED = 'CALL_VERIFIED',
  CONTRACT_GENERATED = 'CONTRACT_GENERATED',
  CONTRACT_SIGNED = 'CONTRACT_SIGNED',
  FIELD_VERIFIED = 'FIELD_VERIFIED',
  ADMIN_APPROVED = 'ADMIN_APPROVED',
  DISBURSED = 'DISBURSED',
  REJECTED = 'REJECTED',
  CLOSED = 'CLOSED',
  REJECTED_BY_MANAGER = 'REJECTED_BY_MANAGER',
  REJECTED_BY_ADMIN = 'REJECTED_BY_ADMIN',
}